+++ activerecord/lib/active_record/store.rb
@@ -1,4 +1,6 @@
+require 'active_support/concern'
 require 'active_support/core_ext/hash/indifferent_access'
+require 'active_support/core_ext/class/attribute'
 
 module ActiveRecord
   # Store gives you a thin wrapper around serialize for the purpose of storing hashes in a single column.
@@ -33,8 +35,17 @@ module ActiveRecord
   #   class SuperUser < User
   #     store_accessor :settings, :privileges, :servants
   #   end
+  #
+  # The stored attribute names can be retrieved using +stored_attributes+.
+  #
+  #   User.stored_attributes[:settings] # [:color, :homepage]
   module Store
     extend ActiveSupport::Concern
+
+    included do
+      class_attribute :stored_attributes
+      self.stored_attributes = {}
+    end
 
     module ClassMethods
       def store(store_attribute, options = {})
@@ -43,7 +54,8 @@ module ActiveRecord
       end
 
       def store_accessor(store_attribute, *keys)
-        keys.flatten.each do |key|
+        keys = keys.flatten
+        keys.each do |key|
           define_method("#{key}=") do |value|
             initialize_store_attribute(store_attribute)
             send(store_attribute)[key] = value
@@ -55,6 +67,8 @@ module ActiveRecord
             send(store_attribute)[key]
           end
         end
+
+        self.stored_attributes[store_attribute] = keys
       end
     end