+++ activerecord/CHANGELOG.md
@@ -1,5 +1,17 @@
 ## Rails 4.0.0 (unreleased) ##
 
+*   Added `stored_attributes` hash which contains the attributes stored using
+    ActiveRecord::Store. This allows you to retrieve the list of attributes
+    you've defined.
+
+       class User < ActiveRecord::Base
+         store :settings, accessors: [:color, :homepage]
+       end
+
+       User.stored_attributes[:settings] # [:color, :homepage]
+
+    *Joost Baaij & Carlos Antonio da Silva*
+
 *   `composed_of` was removed. You'll have to write your own accessor
     and mutator methods if you'd like to use value objects to represent some
     portion of your models. So, instead of: